backend/requirements.txt
fastapi[standard]
uvicorn[standard]
sqlmodel
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
pydantic-settings
openai
langchain
langchain-community
langchain-openai
langchain-text-splitters
chromadb
pypdf
python-docx
aiofiles
